+++ src/main/java/org/wso2/identity/password/validator/hibp/HIBPConnectorConfig.java
@@ -1,31 +1,30 @@
 /*
- * Copyright (c) 2022, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2024, WSO2 LLC. (https://www.wso2.com).
  *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
+ * KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations
  * under the License.
  */
 
-package org.wso2.hibp.connector;
+package org.wso2.identity.password.validator.hibp;
 
 import org.wso2.carbon.identity.governance.IdentityGovernanceException;
 import org.wso2.carbon.identity.governance.common.IdentityConnectorConfig;
+import org.wso2.identity.password.validator.hibp.util.Constants;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Properties;
-
-import static org.wso2.hibp.connector.util.Constants.*;
 
 /**
  * HIBP connector config implementation.
@@ -35,7 +34,7 @@ public class HIBPConnectorConfig implements IdentityConnectorConfig {
     @Override
     public String getName() {
 
-        return CONNECTOR_NAME;
+        return Constants.CONNECTOR_NAME;
     }
 
     @Override
@@ -66,8 +65,8 @@ public class HIBPConnectorConfig implements IdentityConnectorConfig {
     public Map<String, String> getPropertyNameMapping() {
 
         Map<String, String> nameMapping = new HashMap<>();
-        nameMapping.put(CONNECTOR_ENABLE, "Enable HaveIBeenPwned pwned password checker");
-        nameMapping.put(CONNECTOR_API_KEY, "HaveIBeenPwned API key");
+        nameMapping.put(Constants.CONNECTOR_ENABLE, "Enable HaveIBeenPwned pwned password checker");
+        nameMapping.put(Constants.CONNECTOR_API_KEY, "HaveIBeenPwned API key");
         return nameMapping;
     }
 
@@ -75,8 +74,9 @@ public class HIBPConnectorConfig implements IdentityConnectorConfig {
     public Map<String, String> getPropertyDescriptionMapping() {
 
         Map<String, String> descriptionMapping = new HashMap<>();
-        descriptionMapping.put(CONNECTOR_ENABLE, "Validate user passwords with HaveIBeenPwned pwned password checker.");
-        descriptionMapping.put(CONNECTOR_API_KEY, "API key for the HaveIBeenPwned service.");
+        descriptionMapping.put(Constants.CONNECTOR_ENABLE, "Validate user passwords with HaveIBeenPwned pwned " +
+                "password checker.");
+        descriptionMapping.put(Constants.CONNECTOR_API_KEY, "API key for the HaveIBeenPwned service.");
         return descriptionMapping;
     }
 
@@ -84,8 +84,8 @@ public class HIBPConnectorConfig implements IdentityConnectorConfig {
     public String[] getPropertyNames() {
 
         return new String[]{
-                CONNECTOR_ENABLE,
-                CONNECTOR_API_KEY
+                Constants.CONNECTOR_ENABLE,
+                Constants.CONNECTOR_API_KEY
         };
     }
 
@@ -93,8 +93,8 @@ public class HIBPConnectorConfig implements IdentityConnectorConfig {
     public Properties getDefaultPropertyValues(String s) throws IdentityGovernanceException {
 
         Map<String, String> defaultProperties = new HashMap<>();
-        defaultProperties.put(CONNECTOR_ENABLE, "false");
-        defaultProperties.put(CONNECTOR_API_KEY, "");
+        defaultProperties.put(Constants.CONNECTOR_ENABLE, "false");
+        defaultProperties.put(Constants.CONNECTOR_API_KEY, "");
         Properties properties = new Properties();
         properties.putAll(defaultProperties);
         return properties;
