+++ src/main/java/org/wso2/identity/password/validator/hibp/internal/HIBPServiceComponent.java
@@ -1,22 +1,22 @@
 /*
- * Copyright (c) 2022, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2024, WSO2 LLC. (https://www.wso2.com).
  *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
+ * KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations
  * under the License.
  */
 
-package org.wso2.hibp.connector.internal;
+package org.wso2.identity.password.validator.hibp.internal;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -26,12 +26,11 @@ import org.osgi.service.component.annotations.*;
 import org.osgi.service.http.HttpService;
 import org.wso2.carbon.identity.governance.IdentityGovernanceService;
 import org.wso2.carbon.identity.governance.common.IdentityConnectorConfig;
-import org.wso2.hibp.connector.HIBPConnectorConfig;
-import org.wso2.hibp.connector.HIBPServlet;
+import org.wso2.identity.password.validator.hibp.HIBPConnectorConfig;
+import org.wso2.identity.password.validator.hibp.HIBPServlet;
+import org.wso2.identity.password.validator.hibp.util.Constants;
 
 import javax.servlet.Servlet;
-
-import static org.wso2.hibp.connector.util.Constants.HIBP_SERVLET_PATH;
 
 /**
  * HIBP service component
@@ -47,10 +46,10 @@ public class HIBPServiceComponent {
     protected void activate(ComponentContext context) {
 
         Servlet commonAuthServlet = new ContextPathServletAdaptor(new HIBPServlet(),
-                HIBP_SERVLET_PATH);
+                Constants.HIBP_SERVLET_PATH);
 
         try {
-            httpService.registerServlet(HIBP_SERVLET_PATH, commonAuthServlet, null, null);
+            httpService.registerServlet(Constants.HIBP_SERVLET_PATH, commonAuthServlet, null, null);
 
             IdentityConnectorConfig connectorConfig = new HIBPConnectorConfig();
             context.getBundleContext().registerService(IdentityConnectorConfig.class, connectorConfig, null);
