+++ src/main/java/org/wso2/identity/password/validator/hibp/util/Utils.java
@@ -1,22 +1,22 @@
 /*
- * Copyright (c) 2022, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2024, WSO2 LLC. (https://www.wso2.com).
  *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
+ * KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations
  * under the License.
  */
 
-package org.wso2.hibp.connector.util;
+package org.wso2.identity.password.validator.hibp.util;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
@@ -31,7 +31,7 @@ import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.http.message.BasicHeader;
 import org.wso2.carbon.identity.application.common.model.Property;
 import org.wso2.carbon.identity.governance.IdentityGovernanceException;
-import org.wso2.hibp.connector.internal.HIBPDataHolder;
+import org.wso2.identity.password.validator.hibp.internal.HIBPDataHolder;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.InputStream;
@@ -40,7 +40,7 @@ import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.util.*;
 
-import static org.wso2.hibp.connector.util.Constants.*;
+import static org.wso2.identity.password.validator.hibp.util.Constants.*;
 
 /**
  * Utils
